Move RE3Callback to callbacks.py to avoid cyclic dependency. Also update example to use MultiCallbacks.
